US119012: refactor: move last access logic out of Data
