+++ model/data.js
@@ -1,6 +1,6 @@
 import { action, autorun, computed, decorate, observable } from 'mobx';
 import {
-	COURSE_OFFERING, CourseLastAccessFilterId,
+	COURSE_OFFERING,
 	OverdueAssignmentsFilterId, RECORD, TiCVsGradesFilterId, USER
 } from '../consts';
 import { fetchCachedChildren, fetchLastSearch } from './lms.js';
@@ -31,8 +31,6 @@ export class Data {
 		this._userDictionary = null;
 
 		// @observables
-		this.selectedLastAccessCategory = new Set();
-		this.selectedGradesCategories = new Set();
 		this.tiCVsGradesQuadrant = 'leftBottom';
 		this.avgTimeInContent = 0;
 		this.avgGrades = 0;
@@ -218,47 +216,6 @@ export class Data {
 		return recordsByUser;
 	}
 
-	get courseLastAccessDates() {
-		// return an array of size 6, each element mapping to a category on the course last access bar chart
-		const dateBucketCounts = [0, 0, 0, 0, 0, 0];
-		const lastAccessDatesArray = this.getRecordsInView(CourseLastAccessFilterId).map(record => [record[RECORD.COURSE_LAST_ACCESS] === null ? -1 : (Date.now() - record[RECORD.COURSE_LAST_ACCESS])]);
-		lastAccessDatesArray.forEach(record => dateBucketCounts[ this._bucketCourseLastAccessDates(record) ]++);
-		return dateBucketCounts;
-	}
-
-	_bucketCourseLastAccessDates(courseLastAccessDateRange) {
-		const fourteenDayMillis = 1209600000;
-		const sevenDayMillis = 604800000;
-		const fiveDayMillis = 432000000;
-		const oneDayMillis = 86400000;
-		if (courseLastAccessDateRange < 0) {
-			return 0;
-		}
-		if (courseLastAccessDateRange >= fourteenDayMillis) {
-			return 1;
-		}
-		if (courseLastAccessDateRange <= oneDayMillis) {
-			return 5;
-		}
-		if (courseLastAccessDateRange <= fiveDayMillis) {
-			return 4;
-		}
-		if (courseLastAccessDateRange <= sevenDayMillis) {
-			return 3;
-		}
-		if (courseLastAccessDateRange <= fourteenDayMillis) {
-			return 2;
-		}
-	}
-
-	addToLastAccessCategory(category) {
-		this.selectedLastAccessCategory.add(category);
-	}
-
-	setLastAccessCategoryEmpty() {
-		this.selectedLastAccessCategory.clear();
-	}
-
 	get tiCVsGrades() {
 		return this.getRecordsInView(TiCVsGradesFilterId)
 			.filter(record => record[RECORD.CURRENT_FINAL_GRADE] !== null && record[RECORD.CURRENT_FINAL_GRADE] !== undefined)
@@ -323,16 +280,11 @@ decorate(Data, {
 	users: computed,
 	userDataForDisplay: computed,
 	usersCountsWithOverdueAssignments: computed,
-	courseLastAccessDates: computed,
 	tiCVsGrades: computed,
 	tiCVsGradesAvgValues: computed,
 	cardFilters: observable,
 	isLoading: observable,
 	tiCVsGradesQuadrant: observable,
-	selectedLastAccessCategory: observable,
-	selectedGradesCategories: observable,
 	onServerDataReload: action,
-	setApplied: action,
-	addToLastAccessCategory: action,
-	setLastAccessCategoryEmpty: action
+	setApplied: action
 });
