+++ test/model/data.test.js
@@ -450,13 +450,6 @@ describe('Data', () => {
 		});
 	});
 
-	describe('courseLastAccess', () => {
-		it('should return the correct current final grade bucket counts', async() => {
-			const expected = [39, 7, 0, 0, 1, 1];
-			expect(sut.courseLastAccessDates.toString()).to.deep.equal(expected.toString());
-		});
-	});
-
 	describe('overdueAssignments', () => {
 		it('should return a number of users with overdue assignments', async() => {
 			const expected = 2;
